Add set-num-workspaces!, turn num-workspaces into a regular variable
Having num-workspaces as a configurable parameter was a problem,
because workspaces had to be initialized after loading the
configuration file (to be able to honor the value configured fo
num-workspaces).  However, user settings would rely on the number of
configured workspaces to exist, but they would only be created _after_
loading the config file.  This leaded to crashes.

A workaround would be adding a guard procedure to the num-workspaces
parameter. That procedure could initialize/resize the workspaces
vector.  However, that would be too costly.  Each call to
num-workspaces would mean checking the workspaces vector size etc.

So, I ended up implemeting a set-num-workspaces! procedure and
num-procedure is now a global regular variable (fast access).

When shrinking the number of workspaces, windows in workspaces to be
removed will be moved to the last workspace.