grow-window-*!: Add `until' keyword parameter
`until' can be either `window' (default) or `pointer'.  The
`grow-window-*!' procedures will increase/decrease the window size
until it reaches either the next window or the pointer position.

`until: pointer` can be used to semi-interactively resize windows.
Note: it can be slightly non-intuitive when the pointer is inside the
window.